# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TripFit API",
    "version": "v0.0.1"
  },
  "servers": [
    {
      "url": "https://api.tripfit.online",
      "description": "Production API Server"
    }
  ],
  "security": [
    {
      "bearer-jwt": []
    }
  ],
...
  "paths": {
    "/api/v1/users/schedule/regular": {
      "get": {
        "tags": [
          "User Schedule"
        ],
        "summary": "정기 일정 목록",
        "description": "목적: 본인 정기 일정 목록을 조회한다.\n\n결과: 생성 시각 오름차순. 오전·오후·저녁 슬롯은 start/end로 계산된 값.\n",
        "operationId": "listRegular",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseRegularScheduleListResponse"
                },
                "example": {
                  "data": {
                    "items": [
                      {
                        "id": "550e8400-e29b-41d4-a716-446655440000",
                        "title": "출근",
                        "daysOfWeek": "MON,TUE,WED,THU,FRI",
                        "startTime": "09:00:00",
                        "endTime": "18:00:00",
                        "morningStatus": "IMPOSSIBLE",
                        "afternoonStatus": "IMPOSSIBLE",
                        "eveningStatus": "POSSIBLE",
                        "maxVacationDays": 2,
                        "vacationApplyPeriod": "ONE_WEEK_BEFORE",
                        "halfVacationAvailable": false,
                        "holidayRest": true
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "액세스 토큰 없음·무효(AUTH_INVALID_TOKEN)·만료(AUTH_EXPIRED)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_EXPIRED",
                  "message": "액세스 토큰이 만료되었습니다."
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User Schedule"
        ],
        "summary": "정기 일정 생성",
        "description": "목적: 매주 반복되는 정기 일정을 추가한다.\n\n호출 시점: 일정 온보딩·마이페이지에서 정기 일정 추가.\n\n전제: startTime/endTime·요일(daysOfWeek) 필수.\n\n결과: 저장된 정기 일정. 슬롯은 start/end로 계산된다. daysOfWeek는 Weekday(MON~SUN) 콤마 CSV.\n\n주의: 첫 정기 일정 생성 시 hasPreSchedule이 true가 된다(GET /auth/me 등으로 재조회).\n",
        "operationId": "createRegular",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRegularScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "생성 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseRegularScheduleResponse"
                },
                "example": {
                  "data": {
                    "id": "550e8400-e29b-41d4-a716-446655440000",
                    "title": "출근",
                    "daysOfWeek": "MON,TUE,WED,THU,FRI",
                    "startTime": "09:00:00",
                    "endTime": "18:00:00",
                    "morningStatus": "IMPOSSIBLE",
                    "afternoonStatus": "IMPOSSIBLE",
                    "eveningStatus": "POSSIBLE",
                    "maxVacationDays": 2,
                    "vacationApplyPeriod": "ONE_WEEK_BEFORE",
                    "halfVacationAvailable": false,
                    "holidayRest": true
                  }
                }
              }
            }
          },
          "400": {
            "description": "요청 값 검증 실패 (INVALID_INPUT)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "INVALID_INPUT",
                  "message": "입력값이 올바르지 않습니다.",
                  "errors": [
                    {
                      "field": "daysOfWeek",
                      "message": "필수 값입니다."
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "액세스 토큰 없음·무효(AUTH_INVALID_TOKEN)·만료(AUTH_EXPIRED)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_EXPIRED",
                  "message": "액세스 토큰이 만료되었습니다."
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/schedule/regular/{id}": {
      "delete": {
        "tags": [
          "User Schedule"
        ],
        "summary": "정기 일정 삭제",
        "description": "목적: 본인 정기 일정을 삭제한다.\n\n호출 시점: 정기 일정 삭제.\n\n전제: 본인 소유 일정 ID.\n\n결과: 204 No Content.\n\n주의: 정기·개인이 모두 0건이 되면 hasPreSchedule이 false가 된다(GET /auth/me 재조회).\n\n주요 에러: REGULAR_SCHEDULE_NOT_FOUND — 없거나 본인 소유가 아님\n",
        "operationId": "deleteRegular",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "삭제 성공(No Content)"
          },
          "401": {
            "description": "액세스 토큰 없음·무효(AUTH_INVALID_TOKEN)·만료(AUTH_EXPIRED)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_EXPIRED",
                  "message": "액세스 토큰이 만료되었습니다."
                }
              }
            }
          },
          "404": {
            "description": "REGULAR_SCHEDULE_NOT_FOUND — 없거나 본인 소유가 아님",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "REGULAR_SCHEDULE_NOT_FOUND",
                  "message": "정기 일정을 찾을 수 없습니다."
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "User Schedule"
        ],
        "summary": "정기 일정 전체 수정",
        "description": "목적: 기존 정기 일정의 제목·요일·시각·연차 설정을 통째로 갱신한다.\n\n호출 시점: 정기 일정 편집 저장.\n\n전제: 본인 소유 일정 ID.\n\n결과: 갱신된 정기 일정. start/end 변경 시 슬롯을 다시 계산한다.\n\n주요 에러: REGULAR_SCHEDULE_NOT_FOUND — 없거나 본인 소유가 아님\n",
        "operationId": "updateRegular",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRegularScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "수정 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseRegularScheduleResponse"
                },
                "example": {
                  "data": {
                    "id": "550e8400-e29b-41d4-a716-446655440000",
                    "title": "출근",
                    "daysOfWeek": "MON,TUE,WED,THU,FRI",
                    "startTime": "09:00:00",
                    "endTime": "18:00:00",
                    "morningStatus": "IMPOSSIBLE",
                    "afternoonStatus": "IMPOSSIBLE",
                    "eveningStatus": "POSSIBLE",
                    "maxVacationDays": 2,
                    "vacationApplyPeriod": "ONE_WEEK_BEFORE",
                    "halfVacationAvailable": false,
                    "holidayRest": true
                  }
                }
              }
            }
          },
          "400": {
            "description": "요청 값 검증 실패 (INVALID_INPUT)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "INVALID_INPUT",
                  "message": "입력값이 올바르지 않습니다.",
                  "errors": [
                    {
                      "field": "daysOfWeek",
                      "message": "필수 값입니다."
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "액세스 토큰 없음·무효(AUTH_INVALID_TOKEN)·만료(AUTH_EXPIRED)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_EXPIRED",
                  "message": "액세스 토큰이 만료되었습니다."
                }
              }
            }
          },
          "404": {
            "description": "REGULAR_SCHEDULE_NOT_FOUND — 없거나 본인 소유가 아님",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "REGULAR_SCHEDULE_NOT_FOUND",
                  "message": "정기 일정을 찾을 수 없습니다."
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/schedule/personal": {
      "get": {
        "tags": [
          "User Schedule"
        ],
        "summary": "개인 일정 조회",
        "description": "목적: 날짜 구간별 개인(예외) 일정을 조회한다.\n\n호출 시점: 개인 일정 편집 화면 진입.\n\n전제: startDate·endDate 필수.\n\n결과: 날짜당 오전·오후·저녁 슬롯과 uncertain 플래그.\n",
        "operationId": "getPersonal",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "조회 시작일(포함)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2026-08-01"
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "조회 종료일(포함)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2026-08-31"
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponsePersonalScheduleResponse"
                },
                "example": {
                  "data": {
                    "items": [
                      {
                        "id": "550e8400-e29b-41d4-a716-446655440000",
                        "scheduleDate": "2026-08-03",
                        "morningStatus": "IMPOSSIBLE",
                        "afternoonStatus": "POSSIBLE",
                        "eveningStatus": "POSSIBLE",
                        "uncertain": false
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "액세스 토큰 없음·무효(AUTH_INVALID_TOKEN)·만료(AUTH_EXPIRED)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_EXPIRED",
                  "message": "액세스 토큰이 만료되었습니다."
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "User Schedule"
        ],
        "summary": "개인 일정 bulk upsert",
        "description": "목적: 여러 날짜의 개인 일정을 한 번에 저장·삭제한다.\n\n호출 시점: 개인 일정 편집 저장.\n\n전제: items(upsert)와 deletedDates(삭제) 중 하나 이상 필요. 같은 날짜가 양쪽에 있으면 안 된다.\n\n결과: 반영 후 해당 구간 개인 일정 목록.\n\n주의: 첫 저장 시 hasPreSchedule true. 전부 삭제 후 정기·개인 0건이면 false(GET /auth/me 재조회).\n\n주요 에러: INVALID_INPUT — items·deletedDates 교집합 날짜 또는 둘 다 비어 있음\n",
        "operationId": "upsertPersonal",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePersonalScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "저장 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponsePersonalScheduleResponse"
                },
                "example": {
                  "data": {
                    "items": [
                      {
                        "id": "550e8400-e29b-41d4-a716-446655440000",
                        "scheduleDate": "2026-08-03",
                        "morningStatus": "IMPOSSIBLE",
                        "afternoonStatus": "POSSIBLE",
                        "eveningStatus": "POSSIBLE",
                        "uncertain": false
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "INVALID_INPUT — 요청 값 검증 실패 또는 items·deletedDates 교집합 날짜·둘 다 비어 있음",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "INVALID_INPUT",
                  "message": "입력값이 올바르지 않습니다."
                }
              }
            }
          },
          "401": {
            "description": "액세스 토큰 없음·무효(AUTH_INVALID_TOKEN)·만료(AUTH_EXPIRED)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_EXPIRED",
                  "message": "액세스 토큰이 만료되었습니다."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateRegularScheduleRequest": {
        "type": "object",
        "description": "정기 일정 생성 요청. POST /users/schedule/regular. start/end 시각으로 슬롯 계산",
        "properties": {
          "title": {
            "type": "string",
            "description": "표시명 (출근·수업·회의 등)",
            "example": "출근",
            "minLength": 1
          },
          "daysOfWeek": {
            "type": [
              "string",
              "null"
            ],
            "description": "반복 요일. Weekday 콤마 구분(MON~SUN). 생략 시 null",
            "example": "MON,TUE,WED,THU,FRI"
          },
          "startTime": {
            "type": "string",
            "description": "시작 시각",
            "example": "09:00:00"
          },
          "endTime": {
            "type": "string",
            "description": "종료 시각",
            "example": "18:00:00"
          },
          "maxVacationDays": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "여행당 최대 연차 일수. 생략 시 2, 허용 0~10",
            "example": 2,
            "maximum": 10,
            "minimum": 0
          },
          "vacationApplyPeriod": {
            "type": [
              "string",
              "null"
            ],
            "description": "연차 신청 가능 시점. 생략 시 null",
            "enum": [
              "ANY",
              "ONE_WEEK_BEFORE",
              "TWO_WEEKS_BEFORE",
              "ONE_MONTH_BEFORE"
            ],
            "example": "ONE_WEEK_BEFORE"
          },
          "halfVacationAvailable": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "반차 사용 가능 여부. 생략 시 false",
            "example": false
          },
          "holidayRest": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "공휴일 휴무 여부. 생략 시 true",
            "example": true
          }
        },
        "required": [
          "endTime",
          "startTime",
          "title"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "API 에러 응답 envelope",
        "properties": {
          "code": {
            "type": "string",
            "description": "에러 코드",
            "example": "AUTH_INVALID_TOKEN"
          },
          "message": {
            "type": "string",
            "description": "사용자용 에러 메시지",
            "example": "유효하지 않은 토큰입니다."
          },
          "errors": {
            "type": [
              "array",
              "null"
            ],
            "description": "필드 검증 오류 목록 (validation 실패 시)",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
...
      "SuccessResponseRegularScheduleListResponse": {
        "type": "object",
        "description": "API 성공 응답 envelope",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RegularScheduleListResponse",
            "description": "응답 본문"
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "에러 메시지 (성공 시 null)"
          },
          "code": {
            "type": [
              "string",
              "null"
            ],
            "description": "에러 코드 (성공 시 null)"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer-jwt": {
        "type": "http",
        "description": "Authorization: Bearer {accessToken}",
        "name": "bearer-jwt",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}